requirements.txt
# Core Orchestration
langgraph>=0.2.28
langchain>=0.3.14
langchain-core>=0.3.29
langchain-google-genai>=2.0.7
langchain-groq>=0.2.3

# Data Engineering
polars>=1.9.0
pyarrow>=18.0.0
duckdb>=1.1.2

# Visualization
plotly>=5.24.1
kaleido==0.2.1

# Frontend
streamlit>=1.39.0

# Configuration
pydantic>=2.9.2
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
PyYAML>=6.0.2

# Logging & Observability
loguru>=0.7.2
langsmith>=0.1.136

# Utilities
tenacity>=9.0.0
rich>=13.9.2

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0